Reorganize login page and signup page files
